+++ XCode/Project/Src/windows_stubs/d3dx9math.h
@@ -1,0 +1,13 @@
+//
+//  d3dx9math.h
+//  BeebEm
+//
+//  Created by Commander Coder on 04/03/2026.
+//
+
+#ifndef d3dx9math_h
+#define d3dx9math_h
+
+
+
+#endif /* d3dx9math_h */